refactor delete allocation
